# src/generated/swagger.json
{
	"openapi": "3.0.0",
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"UserRole": {
				"enum": [
					"superadmin",
					"manager",
					"user"
				],
				"type": "string"
			},
			"Product": {
				"properties": {
					"id": {
						"type": "string"
					},
					"seller": {
						"$ref": "#/components/schemas/User"
					},
					"winner": {
						"$ref": "#/components/schemas/User"
					},
					"name": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"images": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"starting_price": {
						"type": "number",
						"format": "double"
					},
					"current_highest_bid": {
						"type": "number",
						"format": "double"
					},
					"bidding_start_time": {
						"type": "string",
						"format": "date-time"
					},
					"bidding_end_time": {
						"type": "string",
						"format": "date-time"
					},
					"status": {
						"$ref": "#/components/schemas/ProductStatus"
					},
					"created_at": {
						"type": "string",
						"format": "date-time"
					},
					"updated_at": {
						"type": "string",
						"format": "date-time"
					},
					"bids": {
						"items": {
							"$ref": "#/components/schemas/Bid"
						},
						"type": "array"
					}
				},
				"required": [
					"id",
					"seller",
					"name",
					"description",
					"images",
					"starting_price",
					"bidding_start_time",
					"bidding_end_time",
					"status",
					"created_at",
					"updated_at",
					"bids"
				],
				"type": "object",
				"additionalProperties": false
			},
			"User": {
				"properties": {
					"id": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"password_hash": {
						"type": "string"
					},
					"google_id": {
						"type": "string"
					},
					"provider": {
						"type": "string"
					},
					"role": {
						"$ref": "#/components/schemas/UserRole"
					},
					"created_at": {
						"type": "string",
						"format": "date-time"
					},
					"updated_at": {
						"type": "string",
						"format": "date-time"
					},
					"products": {
						"items": {
							"$ref": "#/components/schemas/Product"
						},
						"type": "array"
					},
					"wonProducts": {
						"items": {
							"$ref": "#/components/schemas/Product"
						},
						"type": "array"
					},
					"bids": {
						"items": {
							"$ref": "#/components/schemas/Bid"
						},
						"type": "array"
					}
				},
				"required": [
					"id",
					"name",
					"email",
					"role",
					"created_at",
					"updated_at",
					"products",
					"wonProducts",
					"bids"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Bid": {
				"properties": {
					"id": {
						"type": "string"
					},
					"product": {
						"$ref": "#/components/schemas/Product"
					},
					"bidder": {
						"$ref": "#/components/schemas/User"
					},
					"amount": {
						"type": "number",
						"format": "double"
					},
					"created_at": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"id",
					"product",
					"bidder",
					"amount",
					"created_at"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ProductStatus": {
				"enum": [
					"pending",
					"approved",
					"rejected",
					"sold",
					"expired"
				],
				"type": "string"
			},
			"CreateProductDto": {
				"properties": {
					"name": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"starting_price": {
						"type": "number",
						"format": "double"
					},
					"bidding_start_time": {
						"type": "string"
					},
					"bidding_end_time": {
						"type": "string"
					}
				},
				"required": [
					"name",
					"description",
					"starting_price",
					"bidding_start_time",
					"bidding_end_time"
				],
				"type": "object",
				"additionalProperties": false
			},
			"PlaceBidDto": {
				"properties": {
					"amount": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"amount"
				],
				"type": "object",
				"additionalProperties": false
			},
			"RegisterDto": {
				"properties": {
					"name": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"password": {
						"type": "string"
					}
				},
				"required": [
					"name",
					"email",
					"password"
				],
				"type": "object",
				"additionalProperties": false
			},
			"LoginDto": {
				"properties": {
					"email": {
						"type": "string"
					},
					"password": {
						"type": "string"
					}
				},
				"required": [
					"email",
					"password"
				],
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {
			"jwt": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT"
			}
		}
	},
	"info": {
		"title": "final-demo",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"paths": {
		"/api/users/me": {
			"get": {
				"operationId": "GetMe",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"role": {
											"$ref": "#/components/schemas/UserRole"
										},
										"email": {
											"type": "string"
										},
										"name": {
											"type": "string"
										},
										"id": {
											"type": "string"
										}
									},
									"required": [
										"role",
										"email",
										"name",
										"id"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Users"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": []
			}
		},
		"/api/users/{id}/role": {
			"patch": {
				"operationId": "UpdateRole",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										}
									},
									"required": [
										"message"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Users"
				],
				"security": [
					{
						"jwt": [
							"superadmin"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"role": {
										"$ref": "#/components/schemas/UserRole"
									}
								},
								"required": [
									"role"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/api/products": {
			"get": {
				"operationId": "GetProducts",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/Product"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Products"
				],
				"security": [],
				"parameters": []
			},
			"post": {
				"operationId": "CreateProduct",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Product"
								}
							}
						}
					}
				},
				"tags": [
					"Products"
				],
				"security": [
					{
						"jwt": [
							"user"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateProductDto"
							}
						}
					}
				}
			}
		},
		"/api/products/my": {
			"get": {
				"operationId": "GetMyProducts",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/Product"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Products"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": []
			}
		},
		"/api/products/{id}": {
			"get": {
				"operationId": "GetProduct",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"bids": {
											"items": {
												"$ref": "#/components/schemas/Bid"
											},
											"type": "array"
										},
										"updated_at": {
											"type": "string",
											"format": "date-time"
										},
										"created_at": {
											"type": "string",
											"format": "date-time"
										},
										"status": {
											"$ref": "#/components/schemas/ProductStatus"
										},
										"bidding_end_time": {
											"type": "string",
											"format": "date-time"
										},
										"bidding_start_time": {
											"type": "string",
											"format": "date-time"
										},
										"starting_price": {
											"type": "number",
											"format": "double"
										},
										"images": {
											"items": {
												"type": "string"
											},
											"type": "array"
										},
										"description": {
											"type": "string"
										},
										"name": {
											"type": "string"
										},
										"winner": {
											"$ref": "#/components/schemas/User"
										},
										"seller": {
											"$ref": "#/components/schemas/User"
										},
										"id": {
											"type": "string"
										},
										"current_highest_bid": {
											"type": "number",
											"format": "double"
										}
									},
									"required": [
										"bids",
										"updated_at",
										"created_at",
										"status",
										"bidding_end_time",
										"bidding_start_time",
										"starting_price",
										"images",
										"description",
										"name",
										"seller",
										"id",
										"current_highest_bid"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Products"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/manager/products": {
			"get": {
				"operationId": "GetPendingProducts",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"total": {
											"type": "number",
											"format": "double"
										},
										"data": {
											"items": {
												"$ref": "#/components/schemas/Product"
											},
											"type": "array"
										}
									},
									"required": [
										"total",
										"data"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Manager"
				],
				"security": [
					{
						"jwt": [
							"manager",
							"superadmin"
						]
					}
				],
				"parameters": []
			}
		},
		"/api/manager/products/{id}/status": {
			"patch": {
				"operationId": "UpdateProductStatus",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										}
									},
									"required": [
										"message"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Manager"
				],
				"security": [
					{
						"jwt": [
							"manager",
							"superadmin"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"status": {
										"$ref": "#/components/schemas/ProductStatus"
									}
								},
								"required": [
									"status"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/api/products/{id}/bids": {
			"get": {
				"operationId": "GetBidHistory",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"limit": {
											"type": "number",
											"format": "double"
										},
										"page": {
											"type": "number",
											"format": "double"
										},
										"total": {
											"type": "number",
											"format": "double"
										},
										"data": {
											"items": {
												"properties": {
													"bidder": {
														"properties": {
															"name": {
																"type": "string"
															},
															"id": {
																"type": "string"
															}
														},
														"required": [
															"name",
															"id"
														],
														"type": "object"
													},
													"created_at": {
														"type": "string",
														"format": "date-time"
													},
													"amount": {
														"type": "number",
														"format": "double"
													},
													"id": {
														"type": "string"
													}
												},
												"required": [
													"bidder",
													"created_at",
													"amount",
													"id"
												],
												"type": "object"
											},
											"type": "array"
										}
									},
									"required": [
										"limit",
										"page",
										"total",
										"data"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Bids"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "page",
						"required": false,
						"schema": {
							"default": 1,
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"default": 10,
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"post": {
				"operationId": "PlaceBid",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"amount": {
											"type": "number",
											"format": "double"
										},
										"message": {
											"type": "string"
										}
									},
									"required": [
										"amount",
										"message"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Bids"
				],
				"security": [
					{
						"jwt": [
							"user"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/PlaceBidDto"
							}
						}
					}
				}
			}
		},
		"/api/auth/register": {
			"post": {
				"operationId": "Register",
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"refreshToken": {},
										"accessToken": {},
										"user": {
											"properties": {
												"role": {
													"$ref": "#/components/schemas/UserRole"
												},
												"email": {
													"type": "string"
												},
												"id": {
													"type": "string"
												}
											},
											"required": [
												"role",
												"email",
												"id"
											],
											"type": "object"
										}
									},
									"required": [
										"refreshToken",
										"accessToken",
										"user"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Auth"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/RegisterDto"
							}
						}
					}
				}
			}
		},
		"/api/auth/login": {
			"post": {
				"operationId": "Login",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"refreshToken": {},
										"accessToken": {},
										"user": {
											"properties": {
												"role": {
													"$ref": "#/components/schemas/UserRole"
												},
												"email": {
													"type": "string"
												},
												"id": {
													"type": "string"
												}
											},
											"required": [
												"role",
												"email",
												"id"
											],
											"type": "object"
										}
									},
									"required": [
										"refreshToken",
										"accessToken",
										"user"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Auth"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/LoginDto"
							}
						}
					}
				}
			}
		},
		"/api/auth/refresh": {
			"post": {
				"operationId": "Refresh",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"refreshToken": {},
										"accessToken": {},
										"user": {
											"properties": {
												"role": {
													"$ref": "#/components/schemas/UserRole"
												},
												"email": {
													"type": "string"
												},
												"id": {
													"type": "string"
												}
											},
											"required": [
												"role",
												"email",
												"id"
											],
											"type": "object"
										}
									},
									"required": [
										"refreshToken",
										"accessToken",
										"user"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Auth"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"refreshToken": {
										"type": "string"
									}
								},
								"required": [
									"refreshToken"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/api/auth/success": {
			"get": {
				"operationId": "GoogleSuccess",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"refreshToken": {
											"type": "string"
										},
										"accessToken": {
											"type": "string"
										},
										"message": {
											"type": "string"
										}
									},
									"required": [
										"refreshToken",
										"accessToken",
										"message"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Auth"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "accessToken",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "refreshToken",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}